Correct minor mistakes and nitpicks
The MSDOS part is just not true, linus specifically mentions Minix as both the inspiration and the insatisfaction twice in his original mail (see https://fossbytes.com/wp-content/uploads/2016/08/linus-torvald-first-linux-email.png). I'm not even sure MS DOS was usable on 386/486 (linus's system).

Concerning the UNIX-like changes, nowadays no system even resembles UNIX anymore. Unix was started in 1969, and even ten years before linux (in 81) it was already starting to show its age. There were several paradigm shifts between unix and linux, and even the first linux kernel version is just not a UNIX-like system but is very much a POSIX-compliant design. You mean POSIX-compliant systems, the set of principles and design standards behind UNIX. Plus, it's better to focus on something students can learn (POSIX) instead of a dead system. Kind of a personal nitpick, feel free to remove the edits.

Also, Linux is not a concatenation of Linus and UNIX but a mistake from a dyslexic colleague when Linus asked him to publish the files on his FTP server. He put an x instead of an s, and it stayed. Linus never wanted his name in the project, in fact he was explicitely against and wanted to name it something else entirely but I can't remember what.

Corrected the part about intel processors only and x86 support which is now Intel and AMD using x86-64/amd64.